Moving dark mode switcher
